requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.7.0
pillow>=10.3.0
python-dateutil>=2.9.0
Levenshtein>=0.25.1
pytest>=8.2.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
pdfplumber>=0.10.0
reportlab>=4.0.0
# optional: pytesseract>=0.3.10  (requires Tesseract binary installed separately)
